Create sol472.java
Pretty cool solution to 472. uses recursion to find "concatenated words" which are words within a list of words that can be made up of other words. the simplest example is a list "dog" "cat" and "catdog" in this case, catdog is a concatenated word, and so would catdogdogcat. This passes that and other more complicated testcases